+++ encoding/config_test.go
@@ -18,7 +18,7 @@ func TestTxEncoding(t *testing.T) {
 	addr, key := tests.NewAddrKey()
 	signer := tests.NewSigner(key)
 
-	msg := evmtypes.NewTxContract(big.NewInt(1), 1, big.NewInt(10), 100000, big.NewInt(1), []byte{}, nil)
+	msg := evmtypes.NewTxContract(big.NewInt(1), 1, big.NewInt(10), 100000, nil, big.NewInt(1), big.NewInt(1), []byte{}, nil)
 	msg.From = addr.Hex()
 
 	ethSigner := ethtypes.LatestSignerForChainID(big.NewInt(1))